+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/activities/ButtonActivity.kt
@@ -2,9 +2,7 @@ package br.com.stone.emeraldcomponentsandroid.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import br.com.stone.emeraldcomponents.basic.EmeraldButton
 import br.com.stone.emeraldcomponentsandroid.R
-import kotlinx.android.synthetic.main.activity_emerald_button.*
 
 class ButtonActivity : AppCompatActivity() {
 
@@ -13,5 +11,4 @@ class ButtonActivity : AppCompatActivity() {
         setContentView(R.layout.activity_emerald_button)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
     }
-
 }