+++ CHANGELOG.md
@@ -1,25 +1,68 @@
 # CHANGELOG
 
-## Unreleased
+## 0.3.5 - 2018-12-13
+
+### Modified
+- [Add option to round emerald button][issue-68]
+
+### Fixed
+- [Fix crash on api 17 on neutral][issue-66]
+ 
+
+## 0.3.4 - 2018-12-13
+
+### Fixed
+- [NumberFormatException on EmeraldMaskedEditText][issue-58]
+- [Fix crash on EmeraldEditText Password toggle][issue-60]
+
+
+### Modified
+- [Changed primary color from #95C93D to #0DB14B][issue-61]
+
+## 0.3.3 - 2018-12-10
+
+### Modified
+- [Added button customizations][issue-52]
+
+## 0.3.2 - 2018-11-30
+
+### Modified
 - Add setCustomColor method to label component 
 
+## 0.3.1 - 2018-11-27
+
+### Fixed
+- Fix label background and border colors when no other color is being set
+
 ## 0.3.0 - 2018-11-06
+
+### Added
 - [Add emerald colors][issue-25]
 - New label component
 
 ## 0.2.3 - 2018-10-01
+
+### Modified
 - [Added Facebook lib to take screenshots][issue-33]
+
+### Modified
 - [Abstract the usage of the screenshot lib][issue-34]
 
 ## 0.2.2 - 2018-09-20
+
+### Modified
 - [Added DateEventList IU tests][issue-27]
 
 ## 0.2.1 - 2018-09-20
+
+### Modified
 - [Added setOnEditActionDone method to EmeraldBaseEditText][issue-20]
 - [Added inputs (EmeraldEditTexts) components IO tests][issue-22]
 - [Added EmeraldDateSelector UI and EmeraldDateFilter tests][issue-24]
 
 ## 0.2.0 - 2018-09-06
+
+### Added
 - [Added LifecycleOwnerConstraintLayout][issue-14]
 
 ## 0.1.1 - 2018-09-06
@@ -38,4 +81,10 @@
 [issue-33]:https://github.com/stone-payments/emerald-components-android/issues/33
 [issue-34]:https://github.com/stone-payments/emerald-components-android/issues/34
 [issue-25]:https://github.com/stone-payments/emerald-components-android/issues/25
+[issue-52]:https://github.com/stone-payments/emerald-components-android/issues/52
+[issue-58]:https://github.com/stone-payments/emerald-components-android/issues/58
+[issue-60]:https://github.com/stone-payments/emerald-components-android/issues/60
+[issue-61]:https://github.com/stone-payments/emerald-components-android/issues/61
+[issue-66]:https://github.com/stone-payments/emerald-components-android/issues/66
+[issue-68]:https://github.com/stone-payments/emerald-components-android/issues/68